+++ src/utils/mod.rs
@@ -47,5 +47,8 @@ pub use json_rpc_error::*;
 mod error_sanitization;
 pub use error_sanitization::*;
 
+mod url;
+pub use url::*;
+
 #[cfg(test)]
 pub mod mocks;
